Move all wikitext-related models/mappers to its own package
