+++ tools/src/test/java/io/github/ygojson/tools/dataprovider/impl/yugipedia/mapper/YugipediaPrintMapperUnitTest.java
@@ -15,7 +15,7 @@ import org.mapstruct.factory.Mappers;
 
 import io.github.ygojson.model.data.Print;
 import io.github.ygojson.model.data.definition.localization.Language;
-import io.github.ygojson.tools.dataprovider.impl.yugipedia.model.CardTable2;
+import io.github.ygojson.tools.dataprovider.impl.yugipedia.model.wikitext.CardTable2;
 import io.github.ygojson.tools.dataprovider.test.CardTable2Mother;
 
 class YugipediaPrintMapperUnitTest {
