+++ tools/src/test/java/io/github/ygojson/tools/dataprovider/impl/yugipedia/mapper/wikitext/CardTable2MapperUnitTest.java
@@ -1,4 +1,4 @@
-package io.github.ygojson.tools.dataprovider.impl.yugipedia.mapper;
+package io.github.ygojson.tools.dataprovider.impl.yugipedia.mapper.wikitext;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,8 +11,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.mapstruct.factory.Mappers;
 
 import io.github.ygojson.tools.dataprovider.impl.yugipedia.YugipediaTestData;
-import io.github.ygojson.tools.dataprovider.impl.yugipedia.mapper.wikitext.WikitextTemplateMapper;
-import io.github.ygojson.tools.dataprovider.impl.yugipedia.model.CardTable2;
+import io.github.ygojson.tools.dataprovider.impl.yugipedia.model.wikitext.CardTable2;
 import io.github.ygojson.tools.dataprovider.test.ModelTestUtils;
 
 class CardTable2MapperUnitTest {
@@ -36,9 +35,8 @@ class CardTable2MapperUnitTest {
 	) {
 		// given
 		final String wikitext = wikitextTestData.wikitext();
-		final Map<String, String> wikitextMap = WIKITEXT_MAPPER.mapCardTable2Template(
-			wikitext
-		);
+		final Map<String, String> wikitextMap =
+			WIKITEXT_MAPPER.mapCardTable2Template(wikitext);
 		// when
 		final CardTable2 cardTable2 = MAPPER.mapToCardTable2(wikitextMap);
 		// then
