+++ tools/src/test/java/io/github/ygojson/tools/dataprovider/impl/yugipedia/mapper/wikitext/InfoboxSetMapperUnitTest.java
@@ -1,4 +1,4 @@
-package io.github.ygojson.tools.dataprovider.impl.yugipedia.mapper;
+package io.github.ygojson.tools.dataprovider.impl.yugipedia.mapper.wikitext;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,8 +11,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.mapstruct.factory.Mappers;
 
 import io.github.ygojson.tools.dataprovider.impl.yugipedia.YugipediaTestData;
-import io.github.ygojson.tools.dataprovider.impl.yugipedia.mapper.wikitext.WikitextTemplateMapper;
-import io.github.ygojson.tools.dataprovider.impl.yugipedia.model.InfoboxSet;
+import io.github.ygojson.tools.dataprovider.impl.yugipedia.model.wikitext.InfoboxSet;
 import io.github.ygojson.tools.dataprovider.test.ModelTestUtils;
 
 public class InfoboxSetMapperUnitTest {
@@ -32,9 +31,8 @@ public class InfoboxSetMapperUnitTest {
 	) throws JsonProcessingException {
 		// given
 		final String wikitext = wikitextTestData.wikitext();
-		final Map<String, String> wikitextMap = WIKITEXT_MAPPER.mapInfoboxSetTemplate(
-			wikitext
-		);
+		final Map<String, String> wikitextMap =
+			WIKITEXT_MAPPER.mapInfoboxSetTemplate(wikitext);
 		// when
 		final InfoboxSet infoboxSet = MAPPER.mapToInfoboxSet(wikitextMap);
 		// then
