+++ tools/src/test/java/io/github/ygojson/tools/dataprovider/impl/yugipedia/model/MarkupStringUnitTest.java
@@ -10,6 +10,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import io.github.ygojson.tools.dataprovider.impl.yugipedia.model.wikitext.MarkupString;
+
 class MarkupStringUnitTest {
 
 	static Stream<Arguments> consideredMarkups() {
